rename all to index and index to core
