--- src/all.ts
@@ -1,7 +1,0 @@
-export * from './boolean'
-export * from './console'
-export * from './json'
-export * from './math'
-export * from './number'
-export * from './string'
-export * from './index'
