+++ src/core.ts
@@ -1,0 +1,1 @@
+export { reactify, reactifyObject, get, set, when, MaybeRef } from '@vueuse/shared'
