+++ src/index.ts
@@ -1,1 +1,7 @@
-export { reactify, reactifyObject, get, set, when, MaybeRef } from '@vueuse/shared'
+export * from './boolean'
+export * from './console'
+export * from './json'
+export * from './math'
+export * from './number'
+export * from './string'
+export * from './core'
